Fix a few small glitches
FlyingTurtle may have collided with the blocks while flying. Now - it's
impossible.
Fixed a bug where the edge of the plant was visible when the plant
stopped moving due to the player being too close.
FireRocket now disappears when it hits a wall.
From now, the plant waits a little longer for it to come out of the
tube.
Turtle's falling speed have been too slow. Now - has been increased.